feat: Node Quarantine support, part 1
See FORGE-16

This implements a first stab at being able to set a managed host as
quarantined, but does not actually implement the quarantine yet. (It
would be easiest to get this part deployed before testing the "actual"
quarantining on a test DPU in a dev environment.)

This change implements support for setting/clearing host quarantine via
gRPC, admin-CLI, and the web UI.

When quarantine is set, the managed host network config version will be
bumped, and set with a quarantine state. The idea is that the DPU will
see the new config version on the next agent check-in, and adjust the
iptables rules to prevent the host from sending any traffic. (The
implementation of this is not part of this MR.)

See merge request nvmetal/carbide!3819