fix: use sgdisk for correcting GPT disk labels
There is no compatibility guarantee between newer versions of
parted and older versions.

parted is used to fix the GPT metadata as well as move the backup
GPT header to the end of the disk.

qcow-imager is based on Debian bookworm which uses parted 3.5.
Images based on Ubuntu 2404 use parted 3.6 to create the qcow2 image.

When an older version of parted encounters metadata from a newer version,
the --fix will fail. When the instance reboots, the boot process
will hang where the UEFI firmware is looking for the shimx.efi file

gdisk reads the entire disk and the primary GPT header to compute the size and where
the secondary GPT header belongs.  This is independent of how the original GPT
label was written

See merge request nvmetal/carbide!3836