chore: Add PreventAllocation classification on TOR peering errors only in case...
chore: Add PreventAllocation classification on TOR peering errors only in case both ports are unhealthy

Due to redundancy, not both ports need to be functioning.
If a single port is unhealthy we still continue to show the `BgpPeeringTor` alert, but don't prevent allocating the host.

See merge request nvmetal/carbide!3847