FORGE-5322: check for a matching sku even when ignore_unassigned_machines to...
Attempts to re-match machines against a sku on an interval. This allows existing sites to auto-assign SKUs to machines but not force machines into the waiting state. When `ignore_unassigned_machines` is true, the state machine will allow machines to go to the "Ready" state without a SKU, but will periodically check if there is a SKU that matches the machine. This allows the BOM validation feature to be enabled on existing sites without interfering with normal operation while automating SKU assignments without user interaction

~~Draft: testing ongoing...~~ done

See merge request nvmetal/carbide!3817