chore: add concurrency option to forge-hw-health
The number of machines scraped concurrently in `forge-hw-health` was recently raised to 16, a level chosen based on testing done without the new `ForgeApiClient`. Since we might benefit from a higher level of concurrency with the new client type, this MR adds a `--concurrency` option to facilitate that testing. Its value can be a positive integer, 'default', or 'machine_count'.

See merge request nvmetal/carbide!3854