fix: move the logic to power cycle the host after toggling the mode on a DPU
fix: move the logic to power cycle the host after toggling the mode on a DPU to the same place site explorer checks whether to create a host or not based on the number of discovered DPUs. Otherwise, the power cycle code was never being reached.

See merge request nvmetal/carbide!3873