feat: add force-restart verification in the state machine
JIRA: https://jirasw.nvidia.com/browse/FORGE-4412

This MR enhances the state machine's ability to verify if a `ForceRestart` was actually initiated by checking the BMC logs of the machine for specific vendor phrases and generic keywords. If the following are found in the logs after a `ForceRestart` is issued, the state machine proceeds. If not, another `ForceRestart` is initiated.

| Keyword | Machine Type |
|---------|--------------|
| `"Server reset."` | HPE |
| `"Server power restored."` | HPE |
| `"The server is restarted by chassis control command."` | Lenovo |
| `"DPU Warm Reset"` | Bluefield |
| `"BMC IP Address Deleted"` | Bluefield |
| `"reset", "reboot", "restart", "power", "start"` | Generic |

During each state transition attempt, the restart verification function checks the managed host's `MachineLastRebootRequested` property to see if the newly introduced `restart_verified` field is set to `false`.

1. If the restart is not verified:
   * The function examines logs from the `requested_timestamp` onward to check if a restart occurred.
   * If a restart is found, the `restart_verified` field is set to `true`, allowing the state transition to proceed.
   * If no restart is found, the new `verification_attempts` field is incremented, and the state machine enters `Wait` to retry up to **two times**.
2. If the maximum attempts (2) are exceeded:
   * The function issues a **second restart** and moves on to the next state.

A new Redfish call for checking BMC logs is introduced in this [libredfish MR](https://gitlab-master.nvidia.com/nvmetal/libredfish/-/merge_requests/295). This verification feature is currently supported only for Lenovo, HPE, and Bluefield machines, as detailed in the libredfish MR.

See merge request nvmetal/carbide!3815