feat: Persistence support in machine-a-tron
Currently, any time you restart machine-a-tron (or re-deploy it if it's
running in k8s), it would generate new machine info (TPM certs, etc) for
the same MAC addresses, resulting in server errors due to duplicate
keys. It effectively meant machine-a-tron was one-shot and you'd have to
wipe the database inbetween runs.

This makes it so you can quit machine-a-tron and re-launch it, or
redeploy it, and it will use the same machine metadata (tpm certs,
various UUID's) as the previous run, which allows machine-a-tron to
simply resume where it left off (as if a physical machine was powered
off and started up again), reporting the same data to the server.

To do this, we simply persist JSON for the running machines to a
configurable directory. We could use a postgres database or something
like that in the future if it gets any more complicated. A followup MR
to forged will implement a simple stateful volume for storing the json
between runs (confirmed to work in local development.)

See merge request nvmetal/carbide!3886