FORGE-5874: remove use of product name in nic fw updates
This removes the use of `product_name` when trying to determine if a DPU needs an update.  it removes the major number from the DPUs firmware version and compares it against a constant.

The constant will need to be updated when a new BFB is released.

See merge request nvmetal/carbide!3888