feat: split up WaitingForPlatformConfiguration into two steps: configure the bios attributes and then setting the boot order.
feat: split up WaitingForPlatformConfiguration into two steps for Vikings: configure the bios attributes and then setting the boot order.

For most vendors, we have to first configure the BIOS to enable IPv4HTTP. After the BIOS config changes take effect, the server will be able to find the corresponding boot option that we want to configure. Also, the redfish NIC info may not valid during the UEFI booting because the redfish info is filled by successful UEFI booting

Dell is the exception to this because we are able to configure the boot order in the call to modify the BIOS attributes.

See merge request nvmetal/carbide!3910