perf: Stop using CTE's for machine snapshot queries
Queries for machine snapshots and managed host snapshots were using
CTE's (Common Table Expressions, "WITH" queries) to clean up and
de-duplicate a lot of things that would otherwise have to be subselects.
This was to avoid needing to join related data in our rust code and
instead get things like hosts, DPU's, interfaces, history, etc all in
one query.

An issue with CTE's is that they are optimization fences, and postgres
will materialize them and scan/project all the data inside them, even
though only a small subset of the data is going to be selected from the
surrounding query. This was causing postgres to read the entire machine
history table into a huge JSON array once per host, even if you're only
asking for a single host's snapshot.

The CTE perf issues seems to be pretty well-documented... in my initial
testing when doing the CTE's, the query times marginally increased,
although the total request time was the same as when we were joining
data in rust.

The fix: Migrate the queries to using subselects which include their own
WHERE clause to limit the data to the requested ID's. Also, add more
relevant indexes. The subselects have the unfortunate tradeoff of
needing to duplicate some SQL in the managed hosts queries, since both
hosts and their attached DPU's need the same joined data. But the
performance is a *lot* better: An EXPLAIN ANALYZE on a managed host
snapshot query for a single host shows the total query time going from
200-300ms down to 2-3ms, on an import of az51's data (over 700 managed
hosts.)

See merge request nvmetal/carbide!3935