feat: Allow to view Machine State History for deleted Machines
When a Machine was force-deleted, there was no way so far to determine the previous history of the Machine via APIs and view it on any tool.

With this change, the Machine history now gets exposed via an additional `FindMachineStateHistories` gRPC endpoint. The function can be invoked even for deleted Machines.

A new carbide-web URL `/admin/:machine_id/state-history` will show the state history independent of whether a Machine exists or not.

See merge request nvmetal/carbide!3942