Fix issue with integration tests running forever
With the machine state controller interval at 1s, and the api refresh
interval at 2s, we were hitting a point where we were waiting on a DPU
to reach hostinit/waitingfordiscovery, but it left the state before we
could check the API to observe it. Fix this by making the refresh
interval 500ms, or half the machine state controller refresh interval.

Also, introduce a timeout for wait_until_machine_up_with_api_state, so
that we don't run forever if this happens again.

See merge request nvmetal/carbide!3955