Decrease stack space used by site explorer fixtures
Per the [slack discussion][0], we're very close to the 2MB default stack
limit, and some new tests are causing us to overflow. A lot of the stack
size is being eaten up by async code (awaiting futures seems to use a
lot of stack space). Placing certain highly-used futures on the heap via
a Box seems to be freeing enough stack space for us to not need to
increase the limit yet.

To do this I had to make TestEnv Send+Sync so that it can be boxed...
which means getting rid of RefCell and using a proper Mutex instead.

[0]: https://nvidia.slack.com/archives/C02RKLCN8BT/p1745455867937049

See merge request nvmetal/carbide!3965