chore: Use PgConnection everywhere instead of Transaction
Callers should have the option of passing either a transaction or just a
simple connection, based on the use case... callers shouldn't be required to
begin a transaction in order to call a database function.

Transactions can be passed as a PgConnection via DerefMut, so this is
fully compatible with either approach.

This should allow us to experiment with not holding a transaction open
for the duration of a MachineStateController run, for instance.

See merge request nvmetal/carbide!3959