Attempt to fix the uefi certs deleted by upgrading the dpu
Recently an issue was uncovered [1] that does not allow the ATF/UEFI
to be properly loaded on upgrade because the certs for the uefi db
were deleted. As per [1] the bug is known and deleted the certs likely
in the upgrade to 2.7.0, but we did not figure out the issue until
attempting to load 2.9.2 update. This commit attempts to re-add the
certs needed to upgrade to 2.9.2, and is being added separately in
order to verify. Once this is validated in dev3, the 2.9.2 can be
un-reverted.

[1] https://nvbugspro.nvidia.com/bug/5230254

See merge request nvmetal/carbide!3990