chore: Remove legacy maintenance mode in database
Maintenance mode has been gradually replaced with health alerts over the last releases:
- Setting a maintenance mode also triggered an associated health alert being set on the Machine
- Host updates and DPU reprovisioning did not use maintenance mode anymore but more specialized health reports

However the SetMaintenance APIs and associated maintenance_reference fields persisted, and the data was stored in a separate location. This lead to various sync conflicts between the Maintenance Request and the Health Override.

With this change, the legacy storage location for Maintenance Mode data is removed, and therefore also the requirement of syncing data is removed.

The Maintenance Mode APIs are however still available:
- Setting Maintenance Mode will continue placing a health alert on the Host
- The `maintenance_reference` message on the field will reflect the message of the `Maintenance` health report.

See merge request nvmetal/carbide!3987