fix: admin cli vpc-peering create use positional argument
Update the `vpc-peering create` command syntax from:
```
./forge-admin-cli vpc-peering create --vpc1-id <vpc1_id> --vpc2-id <vpc2_id>
```
to:
```
./forge-admin-cli vpc-peering create <vpc_id> <vpc_id>
```

With the previous approach, user can be confused about the ordering of vpc1-id and vpc2-id not preserved in the output like below
```
$ ./forge-admin-cli vpc-peering create --vpc1-id bec60961-3133-4964-9511-22c9bddb1baa --vpc2-id 381468a8-33b5-44ad-a5de-d29c2bc093ed
+--------------------------------------+--------------------------------------+--------------------------------------+
| Id                                   | VPC1 ID                              | VPC2 ID                              |
+======================================+======================================+======================================+
| d35057d0-2d9b-496e-bf7d-d5ac54d90252 | 381468a8-33b5-44ad-a5de-d29c2bc093ed | bec60961-3133-4964-9511-22c9bddb1baa |
+--------------------------------------+--------------------------------------+--------------------------------------+
```
This happens because the VPC IDs are stored alphabetically in the database, rather than in the order provided by the user. To eliminate this confusion, we have removed the `--vpc1-id` and `--vpc2-id` parameter names. Users now provide the VPC IDs as positional arguments instead.

See merge request nvmetal/carbide!3993