Machine lifecycle test - Add firmware downgrade step
Adding a step to downgrade the DPU firmware (BMC and CEC + BFB) to DOCA 2.5.0 before continuing to factory reset and force-delete the machine as normal.

This will test the firmware auto-upgrade during Forge ingestion, as the DPUs we receive for Azure sites are at 2.5.0.

New flag to control whether the downgrade is enabled in a GitLab pipeline: \`FW_DOWNGRADE\`

See merge request nvmetal/carbide!3547