GB200 scout image fixes
Pushing the changes for GB200, with the changes below we were able to fully discover and ingest some GB200 nodes.
Switched the scout image to the nvidia 64K page kernel, and made sure the kernel is decompressed before mkosi puts it into the image.
Added an extra glob to the network interface match, since this system shows them as enP not enp.

See merge request nvmetal/carbide!4009