machine-a-tron: Properly emulate Host-to-DPU DHCP behavior
Improve the code that emulates the DPU-responding-to-host-DHCP behavior,
to actually be a real request channel between the mock Host and the mock
DPU. This decreases the number of API calls significantly, since we no
longer need to "cheat" and get configuration from the server in order to
satisfy a host DHCP request: instead the mock DPU can just respond to
the request with its ManagedHostNetworkConfig that it already has in
scope. This also makes it easy to simply let the HostMachine boot as
early as it wants (not having to artificially wait for DPU's to be
ready), because when it sends the DHCP request, the request will sit in
a channel until the DPU boots up, gets its network config, and replies
to it. This should make the process to getting to Ready even faster
because we don't have to sleep 5 seconds repeatedly until the DPU's all
come up.

Also, for other DHCP's (BMC addresses, and the DPU itself), the existing
dhcp_relay code was intended to actually support listening on a UDP port
for real dhcp requests, but this code has been broken for at nearly a
year or so. Using the API is the only way that works today, so just
delete dhcp_relay and fully depend on using the API (this simplifies
things a lot.)

See merge request nvmetal/carbide!4011