Fix flaky test_metrics_integration test
We still seem to miss the WaitingForDiscovery state, possibly due to
throttling the API calls to check the state. But instead we can just
wait for the Ready state, which takes a bit longer but we're guaranteed
not to "miss" it, because it stays at that state.

See merge request nvmetal/carbide!4049