Fixing the kernel at 6.8.0-45 as the latest doesn't seem to work on some Dell servers
Several people asked me to look at a Dell that was stuck loading scout.    Srikar also provided a log showing that the system was hitting a general protection fault as the kernel was starting up.    Given that I have other images which are working, I'm downgrading the kernel used in scout to a known working one.

See merge request nvmetal/carbide!4080