MLT: Add config for downgrading to DOCA 2.2
For HFA site bring-up we need to test ingestion of DOCA 2.2 DPUs.

Adding config for testing both 2.2.0 and 2.2.1 as we don't yet know which version is installed there.

We have been told there are Bluefield-3s with DOCA 2.2.0 but the documentation says this is an unsupported combination. For now we will test and see.

See merge request nvmetal/carbide!4085