Use wget to track timestamps on files that do not change often
The dpu build has a lot of files it downloads and those do not necessarily change. Some of them were previously curl'd into place and it would cause .1, .2, .3 for each download. Using wget -N -P will track the change of the file and not download it again. This will not affect our CI Builds because those all are done one time in a docker container.

Oh, also, cleaned up some files we dont use anymore in the build.

https://jirasw.nvidia.com/browse/FORGE-6211

See merge request nvmetal/carbide!4094