fix(dpu-agent): Add an 'apt-get autoclean' step to self-upgrade
As of !3906 we have a daily cleanup job that does an `apt clean`, but
it's probably a better practice to clean up incrementally as we download
each new forge-dpu version.

The `autoclean` operation just removes any cached packages that are
no longer referenced by any repository source, so we can just run this
after the `update` operation and we know the new contents of the repo.

See merge request nvmetal/carbide!4109