feat: Host quarantine support in forge-dpu-agent
Per FORGE-16: If the managed host is in a quarantined state, write an
nvue config that blocks all ipv4 and ipv6 traffic, ingress and egress.

Tested on a DPU in dev3 in eth virtualization mode (with nvue), fnn was
not tested explicitly but should work, since the ACL format should be
the same.

See merge request nvmetal/carbide!3905