Add a message for the CLI when the client certificate is expired
Right now, when there's an issue with the client certificate (usually expiration), the TLS client will retry in an error loop, with a vague message about not being able to connect. This has lead to a lot of folks (including me) getting lost as to why forge-admin-cli wasn't working. This change updates the TLS client to return a TLS configuration error when there is a client certificate chain. This error is treated as "permanent" by the calling code. I also added a logging message, to make it more clear to interactive callers what the issue likely is. 

Note this may have an issue in other uses of the Forge TLS client, since I'm not sure where they're getting the client certificate chains, I'm not sure (hoping to get some advice in this MR).

Before:

```sh
$ forge-admin-cli v                                                                                                                                                                                                                                                                                                                                                                                                                                               (base)
2025-05-23T00:22:35.421863Z ERROR rpc::forge_tls_client: error connecting client to forge api (url: https://carbide-api.forge), will retry: status: Unknown, message: "client error (SendRequest)", details: [], metadata: MetadataMap { headers: {} }

<retry loop>
```

After:

```sh
$ forge-admin-cli v                                                                                                                                                                                                                                                                                                                                                                                                                                                    (base)
2025-05-23T01:36:02.487531Z ERROR rpc::forge_tls_client: Client certificate is expired, perhaps you need to run nvinit again?
Error:
   0: The API call to the Forge API server returned status: Unavailable, message: "Configuration error: Invalid client cert: invalid peer certificate: Expired", details: [], metadata: MetadataMap { headers: {} }
   1: status: Unavailable, message: "Configuration error: Invalid client cert: invalid peer certificate: Expired", details: [], metadata: MetadataMap { headers: {} }
...
```

Fixes https://jirasw.nvidia.com/browse/FORGE-6223

See merge request nvmetal/carbide!4103