Fix lldp on the DPU
Up until a few releases of doca/hbn ago, lldp was broken. It was commented out. This commit uncomments it, fixes up how it works on the new (and old) lldp formats, and runs a version check in discovery to ensure its not run when it was known broken.

Ref: https://jirasw.nvidia.com/browse/FORGE-6253

See merge request nvmetal/carbide!4114