Test with multiple carbide-api instances in integration tests
I wanted to find whatever bugs happen if we actually run multiple
carbide-api instances, and to my (mostly) surprise, it seems to work ok.
This makes it so our integration tests actually exercise this behavior,
so that we don't add code which relies on their only being one instance.

The integration test emulates a load balancer by customizing the api
client to switch between each server on every API call. (And similarly,
the grpcurl wrapper selects one at random.) We can actually reuse this
in the future if we want to support client-side load balancing, but for
now it's good enough to just make this an integration-test-only feature.

One hiccup is metrics: Each API server has different metrics, so when the
`wait_for_metric_line` helper loops waiting for a metric line to show
up, it rotates between servers on each iteration. In reality, we need to
decide how we want to report things like state iteration metrics if only
one API server is actually performing state controllers at a time.

See merge request nvmetal/carbide!4122