Fix for supermicro grace/grace servers giving a synchronous error when running iPXE
The problem was iPXE trying to stop a driver which was already stopped.   This had been fixed upstream, so I've made that fix a patch and put into our build.   The patch is only applied to the ARM build.

I've tested on a DPU as well as GB200 and Grace/Grace.

See merge request nvmetal/carbide!4128