Add qcow-imager for ARM64
This is mostly a copy of the existing debian based one for x86_64, but I switched the distro to ubuntu and pulled in the nvidia 64K kernel for better compatibility with GB200.
The only changes to the disk_imaging.sh script were to make sure it converts from aarch64 to arm64 for the filenames in the ubuntu/debian repos.
It has been tested in a Grace/Grace and correctly wrote out the image.

See merge request nvmetal/carbide!4130