Removing dkms repo pinning.
The nvidia drivers now have a dependency on the DKMS package in the nvidia repo, previously this version gave us problems, but it appears that it's been fixed.  So we can remove the preferences file which was pinning the DKMS package to the Ubuntu repo.

See merge request nvmetal/carbide!4134