limit error back-off in forge-hw-health to 15 minutes
This MR fixes the problem of "losing" hosts described in FORGE-6301 by limiting the error back-off to 15 minutes instead of up to a whole day. Scrape errors don't affect the scrape interval as much as they used to because 1) scrapes are concurrent, so long error times are less likely to happen one after another, and 2) scrapes don't take as long to fail as they used to now that we have a better gRPC client. This MR also increases the number of concurrent scrapes from 16 to 32.

See merge request nvmetal/carbide!4126