Extract api-test helpers into a new api-test-helper crate
Lay more groundwork for additional usage of this code, particularly for
integration tests for a future rewrite of ssh-console.

Everything but main.rs from api-test/tests/integration becomes
common/api-helper, and main.rs becomes api/tests/integration.rs.
(Reminder that a `tests` folder in the root of a crate already means
"integration tests", and each file in there is compiled as its own
crate, see the [docs][0].) We never needed api-test to be its own crate,
it could have always lived in api's tests directory.)

I could have left `api-test` as a crate and moved everything to `src`,
but I think `api-test-helper` is a better name as it reflects that the
purpose is to help us write other tests, and there are no actual tests
in the crate itself.

[0]: https://doc.rust-lang.org/book/ch11-03-test-organization.html#the-tests-directory

See merge request nvmetal/carbide!4158