Only add a uefi db cert to bfb if none exist
There is bug that causes our dpu installs to get very slow if there is
an existing failed dpu bmc task. It is not easy to have a failed bmc
task, but things that have failed one install and not been depowered
at the host level could have this failed bmc task since we started
adding certs. If a human intervenes to fix a broken dpu, they could
amplify this problem until the dpu has been fully powered off by
powering off the host, which gets rid of failed dpu bmc tasks. Until
the upstream installer can ignore old failed tasks, we will work
around it by only adding the certificate task if we know the
certificate is not in the db. This commit checks to see that the
number of certs is not zero and then adds it if it is zero.

See merge request nvmetal/carbide!4153