Add stub for rewrite of ssh-console in Rust
Currently the goal is to properly integration-test the existing
("legacy") ssh-console, to make sure existing behavior is captured,
before doing the rust rewrite. (I want to put the rust rewrite of
ssh-console in this repo so we can easily call rpc stuff without having
to worry about importing it as a cargo crate, plus all the other
benefits of having integration tests, etc.) These tests spawn a carbide
cluster with two mock machine-a-tron hosts, and verify that it can
connect via ssh-console to each of them, and that the machine ID shows
up in the mocked SSH prompt. ssh-console is added as a submodule under
ssh-console/legacy/ssh-console to facilitate this. It's completely
optional unless you're running these tests. [0]

To do this, machine-a-tron is now running a mock BMC ssh server,
listening on port 2222, leveraging the codepath that does `ip addr add`
when running BMC mocks, so that each BMC mock gets a unique IP to run
SSH on the same port. ssh-console has a recent MR to change the SSH port
as well, to make this work.

The mock SSH server emulates what happens when you connect to a dell
BMC: renders `racadm>>` on the default prompt, and if you run `connect
com2`, it renders a prompt like `root@<machine_id> #` (then drops back
to racadm if ctrl+\ is sent.) This gets ssh-console to issue the correct
state logic on the connection so that we can do assertions: particularly
that SSH'ing to <host_id>@localhost or <instance_id>@localhost gives us
a prompt that looks like `root@<machine_id> #`. More assertions will
likely come later as I start to actually write the new rust code.

This change also runs a local instance of ipmi_sim to simulate the
relevant IPMI calls that ssh-console needs. Although there is just a
single ipmi_sim for all mocked hosts at the moment.

[0] The test code that launches ssh-console depends on a working
toolchain for compiling it (it chdir's into the submodule and runs make,
for instance), so the tests are only active if you set the
RUN_SSH_CONSOLE_TESTS environment variable, to avoid slowing anything
down in CI. The submodule and tests should be completely ignored unless
that env var is set when running tests.

See merge request nvmetal/carbide!4175