feat: Show source code reference in Web UI
When the state handler returns with an `Ok` outcome, this change will record the source code file and line number consistently in the database.

The admin web UI can then be used to view where the state handler returned in order to get more information about why a Machine might be stuck due to a certain condition.

See merge request nvmetal/carbide!4195