fix(api): Reject force-delete requests for machines associated with an instance type
When an instance is created in cloud-api, it now sends the instance type ID it thinks/expects the source machine has on the site when it makes its call to Carbide.  If there is a mismatch, then the request is rejected by Carbide.

Force-deleting a machine which deletes the machine record entirely, and that include the instance type details.  Eventually, cloud and site would be consistent due to an inventory sync process, but it would be cloud getting an update to remove the instance type association because the site is the source of truth, and that can mess up allocation relationships. E.g., you could have allocated 10 machines to a tenant, but the force-delete actually removed a machine from an instance-type, and it's no longer possible to allocate 10 machines.

Also, as we begin to move compute allocation data to the site, which is going to start soon, that same issue would then be present on the site.

See merge request nvmetal/carbide!4216