fix: wait for five minutes before transitioning to the BOSS secure erase error handling states
fix: when creating a bios job to secure erase the BOSS drives on a Dell server, it takes sometime between creating the job and scheduling the job. Wait five minutes before transitioning to the error handling state for secure erase to ensure that the job has been scheduled prior to declaring an error.

See merge request nvmetal/carbide!4221