Faster integration tests for ssh-console
Instead of running machine-a-tron and a complete API server from
scratch, run a mock API server that responds to calls the way we expect,
by copying over the forge protos we want to mock, but only defining a
small subset of the calls we want to implement.

This allows us to easily respond to GetBMCMetaData with a response which
sets a custom SSH port, allowing running multiple SSH mocks without
having to do the whole `ip address add` trick to add aliases to
loopback, which should allow tests to run properly in CI.

For legacy tests, simply run against a single mock server so that we can
set a single BMC SSH port override.

This also cherry-picks out some of the test changes from !4226 working
on a minimum viable product of the rust rewrite, to make that MR
smaller.

See merge request nvmetal/carbide!4231