fix(agent): Raised the MAX_EXPECTED_SIZE limit for file reads by a lot
It's currently very possible for a tenant to get the agent into a state where it will fail to update its NVUE config due to a diff check failing, and this can even prevent it from successfully terminating.
Current path that lead to this was a tenant adding a lot of security group rules, but anything that pushes NVUE size up will do it.

```
error="NVUE config at /var/lib/hbn/var/support/nvue_startup.yaml: read_limited /var/lib/hbn/var/support/nvue_startup.yaml: 43234 > 16384 bytes"
```

Additionally, we might want to ignore the diff check that triggered this for NVUE configs when we are switching to the admin network so that we never have to worry about terminations failing because of this.

See merge request nvmetal/carbide!4260