ssh-console: impi support, use single backend for multiple frontends
This implements ipmi-based console support using a PTY to interact with
ipmitool, and implements sharing support so that multiple connections to the
same backend share a single connection. The next step will be implementing the
loki logs, which should be a simple matter of eagerly connecting to each
backend, and attaching a single long-lived observer to each one (via the same
broadcast channel it's already sending to SSH clients) to ship logs off.

See merge request nvmetal/carbide!4248