ssh-console: fix reference cycle preventing frontends from getting disconnected
We have to keep a copy of broadcast_to_frontend_tx in each BackendHandle
so that new frontends can call .subscribe() on it, but we were using a
strong reference, which was causing the channel to never be closed if
the backend disconnected while we still had connected frontends. Fix it
by keeping a WeakSender in BackendHandle instead. That way if the
backend closes, frontends can be disconnected immediately.

See merge request nvmetal/carbide!4279