ssh-console: Less noisy error handling, faster auth rejections
Currently ssh client disconnects and EOF's are logged as scary-looking
errors. To fix this we have to use a custom error type which can be
converted either from a russh error or an eyre error, so that we can dig
in and check if it was a russh error, and if so, downgrade
disconnects/eof's to debug level.

Also, reject auth in 30ms rather than 1s. This corresponds with the
duration it takes to do an auth rejection from carbide, multiplied by a
few times. (The idea is for rejections to be roughly constant time to
avoid timing attacks.)

See merge request nvmetal/carbide!4288