MLT: Disable schedule pipeline on failure
Originally we wanted the scheduled pipeline to continue failing so it would make noise in the Slack channel, however I don't think this is beneficial in its current format given the limitations of the information displayed by the Slack integration.

Instead we will disable the pipelines after they fail and continue reporting the underlying issues in Slack.

NOTE: for this to work for a given pipeline, someone must manually add a variable named PIPELINE_SCHEDULE_ID which is set to the value of that scheduled pipeline's ID. This is a workaround for the fact that Gitlab does not make that variable available to us.

I have set this for the main 3 MLT pipelines.

See merge request nvmetal/carbide!4317