ssh-console: Overhaul/Refactor reconnection logic
While implementing metrics for the new ssh-console, I realized in order
to have stats about which backends are healthy, reconnecting, etc, we
need to overhaul how reconnecting works to not be so dependent upon
console logging.

This overhauls ssh-console to behave more like the legacy ssh-console, where
"connect to every backend" is more core behavior, and not just a side-effect of
logging.

Before: We connected to backends "on-demand", and we only eagerly connected to
all backends if console logging was enabled. The logger itself was responsible
for reconnection logic.

After: Always connect to all backends, and retry forever, even if we're not
logging. Because logging is optional, and even if we are logging, we don't want
a case where the logger is sleeping waiting to retry (due to exponential
backoff) and then a user comes in and triggers an on-demand connection while
we're not logging.

Changes:

- Make a new mod, "backend_session", which is responsible for discovering
  consoles, reconnecting to them, and logging their output
- Rename backend to backend_connection, to reflect that it's responsible for a
  single connection to a backend console, and not retry logic.
- Implement a message relay that echoes "BMC console not connected" back to the user
  if they SSH in and the console is not connected. (This makes it easier for the
  user to see that their connection to ssh-console is working, but the BMC
  itself is not.)

See merge request nvmetal/carbide!4322