chore: Populate Machine::ib_status_observation from IbFabricMonitor
Previously we queried UFM for the latest IB connection status in every Machine state handler iteration. The querying of IB status was however limited to certain Machine lifecycle states.

This change moves the querying of UFM and the update of `ib_status_observation` into the IBFabricMonitor tasks. The benefits of this are:
- We only call the UFM ports API (that doesn't support querying by GUID) once instead of once for every Machine.
- IB status is updated for Machines in **all** states
- Information can be used for continuous reconciliation later on (https://jirasw.nvidia.com/browse/FORGE-6424).
- Information can be looked up across various fabrics, which means we get closer to multi-fabric support.

See merge request nvmetal/carbide!4326