fix: Fix IB partition deletion
The IB partition state handler has a check which queries UFM for the existence of the partition. This check did not interpret the response from UFM correctly as "not found" anymore, since the logic to interpret values of `{}` as not found was removed.. With this change, the UFM response is again interpreted as not found, and partition deletion should succeed.

See merge request nvmetal/carbide!4338