Making the nvidia repo 1st choice, added cuda to scout.
It seems the Ubuntu repo has updated their nvidia packages to 575.64, the nvidia repo is at 575.57.    We have ended up with a mixture of both due to dependencies.

This MR makes the nvidia repo the higher priority, so we end up with a consistent set of versions.

I've tested it on a Viking in QA2, and the dcgmi long test passed with the test image.

I have also moved cuda into scout, rather than being a post-boot download, as that could be unreliable.

See merge request nvmetal/carbide!4331